requirements.txt
# ── core ──────────────────────────────────────────────────────
flask>=3.0.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0

# ── classical ML ───────────────────────────────────────────────
scikit-learn>=1.4.0

# ── deep learning ──────────────────────────────────────────────
tensorflow>=2.16.0        # includes Keras 3

# ── transformers ───────────────────────────────────────────────
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.30.0        # GPU helpers for HF transformers

# ── tokenizers (faster HF tokenisation) ────────────────────────
tokenizers>=0.19.0
sentencepiece>=0.2.0

# ── pattern mining + linguistic features (novel contribution) ─
mlxtend>=0.23.0
vaderSentiment>=3.3.2
textstat>=0.7.3

# ── LLM classifier / generator (novel contribution) ────────────
# Uses ASU Research Computing's OpenAI-compatible LLM endpoint.
# Key + base URL configured in .env (see .env.example).
openai>=1.40.0
python-dotenv>=1.0.0

# ── production WSGI server (Render / Railway) ─────────────────
gunicorn>=21.0.0
